small tweaks to remove compiler warnings, added strlen for null string test changed configure directory search paths to include fink’s use of /sw or soon /opt/sw Corrected calculation of size of string for a base64 encode/decode avoid array overrun crash SSL call ASN1_STRING_data -> ASN1_STRING_get0_data to avoid deprecated call Corrected length for Finder Info encoding and clarified message

+++ contrib/fsspy.c
@@ -208,7 +208,7 @@ fsevent_callback( ConstFSEventStreamRef ref, void *info,
 	event_insert( &event_head, new );
 
 	if ( debug ) {
-	    printf( "%-10llu %-10lu %s\n", eids[ i ], eflags[ i ], paths[ i ] );
+	  printf( "%-10llu %-10lu %s\n", eids[ i ], ( unsigned long )eflags[ i ], paths[ i ] );
 	}
     }
